utils: moved filter_polygons upwards: from coco.py to shapely.py
