+++ sahi/utils/coco.py
@@ -224,28 +224,7 @@ class CocoAnnotation:
         self._shapely_annotation = shapely_annotation
 
     def get_sliced_coco_annotation(self, slice_bbox: List[int]):
-        def filter_polygons(geometry):
-            """
-            Filters out and returns only Polygon or MultiPolygon components of a geometry.
-            If geometry is a Polygon, it converts it into a MultiPolygon.
-            If it's a GeometryCollection, it filters to create a MultiPolygon from any Polygons in the collection.
-            Returns an empty MultiPolygon if no Polygon or MultiPolygon components are found.
-            """
-            if isinstance(geometry, Polygon):
-                return MultiPolygon([geometry])
-            elif isinstance(geometry, MultiPolygon):
-                return geometry
-            elif isinstance(geometry, GeometryCollection):
-                polygons = [geom for geom in geometry.geoms if isinstance(geom, Polygon)]
-                return MultiPolygon(polygons) if polygons else MultiPolygon()
-            return MultiPolygon()
-
         shapely_polygon = box(slice_bbox[0], slice_bbox[1], slice_bbox[2], slice_bbox[3])
-        samp = self._shapely_annotation.multipolygon
-        if not samp.is_valid:
-            valid = make_valid(samp)
-            valid = filter_polygons(valid)
-            self._shapely_annotation.multipolygon = valid
         intersection_shapely_annotation = self._shapely_annotation.get_intersection(shapely_polygon)
         return CocoAnnotation.from_shapely_annotation(
             intersection_shapely_annotation,
